third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnectionIceEvent-constructor-expected.txt

This is a testharness.js-based test.
PASS RTCPeerConnectionIceEvent with no arguments throws TypeError
FAIL RTCPeerConnectionIceEvent with no eventInitDict (default) assert_equals: expected (object) null but got (undefined) undefined
FAIL RTCPeerConnectionIceEvent with empty object as eventInitDict (default) assert_equals: expected (object) null but got (undefined) undefined
PASS RTCPeerConnectionIceEvent.candidate is null when constructed with { candidate: null }
PASS RTCPeerConnectionIceEvent.candidate is null when constructed with { candidate: undefined }
PASS RTCPeerConnectionIceEvent with RTCIceCandidate
PASS RTCPeerConnectionIceEvent with non RTCIceCandidate object throws
PASS RTCPeerConnectionIceEvent bubbles and cancelable
Harness: the test ran to completion.
